Check for New and Missing Members
Feature: The bot can now check if there is a new member in the server. If one or more are detected, they're appended to the member file.

Feature: The bot can now check if a member is missing from the server, and alerts in general if this is the case.

Fix: Changed all instances of .write() to print().